Implementacao de Alert Dialog para remocao de um aluno

+++ formacao-android/Agenda/app/src/main/java/com/example/agenda/ui/activity/util/FilterContextMenu.kt
@@ -1,6 +1,9 @@
 package com.example.agenda.ui.activity.util
 
 
+import android.content.Context
+import android.content.DialogInterface
+import androidx.appcompat.app.AlertDialog
 import com.example.agenda.dao.StudentDAO
 import com.example.agenda.model.Student
 import com.example.agenda.ui.activity.adapter.StudentCustomAdapter
@@ -10,10 +13,18 @@ enum class FilterContextMenu {
         override fun action(
             student: Student,
             studentDAO: StudentDAO,
-            studentCustomAdapter: StudentCustomAdapter
+            studentCustomAdapter: StudentCustomAdapter,
+            context: Context
         ) {
-            studentDAO.remove(student)
-            studentCustomAdapter.remove(student)
+            AlertDialog.Builder(context)
+                .setTitle("REMOVING STUDENT")
+                .setMessage("Do you want to remove this student?")
+                .setPositiveButton("YES") { dialog, which ->
+                    studentDAO.remove(student)
+                    studentCustomAdapter.remove(student)
+                }
+                .setNegativeButton("NO", null)
+                .show()
         }
     },
     ;
@@ -21,6 +32,7 @@ enum class FilterContextMenu {
     abstract fun action(
         student: Student,
         studentDAO: StudentDAO,
-        studentCustomAdapter: StudentCustomAdapter
+        studentCustomAdapter: StudentCustomAdapter,
+        context: Context
     )
 }
